Refactor: Complete Tier 1 component refactoring (Issue #9B)
- Extract DiagnosticsPanel, ColorDetailPanel, TokenInspector
- Create 18 focused subcomponents + 15 custom hooks
- Reduce main files by 75% average
- All TypeScript checks passing
- Follows established modular pattern

Components refactored:
- DiagnosticsPanel: 450 → 95 LOC (79% reduction)
- ColorDetailPanel: 432 → 60 LOC (86% reduction)
- TokenInspector: 358 → 150 LOC (56% reduction)

Total: 1,240 → 305 LOC in main files
