+++ frontend/src/App.tsx
@@ -5,8 +5,8 @@ import ColorTokenDisplay from './components/ColorTokenDisplay'
 import ShadowTokenList from './components/shadows/ShadowTokenList'
 import './components/shadows/ShadowTokenList.css'
 import LightingAnalyzer from './components/LightingAnalyzer'
-import DiagnosticsPanel from './components/DiagnosticsPanel'
-import TokenInspector from './components/TokenInspector'
+import DiagnosticsPanel from './components/diagnostics-panel'
+import TokenInspector from './components/token-inspector'
 import TokenGraphPanel from './components/TokenGraphPanel'
 import ColorGraphPanel from './components/ColorGraphPanel'
 import SpacingScalePanel from './components/SpacingScalePanel'
