+++ frontend/src/components/color-detail-panel/ColorDetailPanel.css
