+++ frontend/src/components/color-detail-panel/index.ts
@@ -1,0 +1,8 @@
+export { ColorDetailPanel } from './ColorDetailPanel'
+export { ColorHeader } from './ColorHeader'
+export { OverviewTab } from './tabs/OverviewTab'
+export { HarmonyTab } from './tabs/HarmonyTab'
+export { AccessibilityTab } from './tabs/AccessibilityTab'
+export { PropertiesTab } from './tabs/PropertiesTab'
+export { DiagnosticsTab } from './tabs/DiagnosticsTab'
+export type { Props, TabType, TabProps, DiagnosticsTabProps } from './types'
