+++ frontend/src/components/color-detail-panel/types.ts
@@ -1,0 +1,18 @@
+import type { ColorToken } from '../types'
+
+export type TabType = 'overview' | 'harmony' | 'accessibility' | 'properties' | 'diagnostics'
+
+export type Props = {
+  color: ColorToken | null
+  debugOverlay?: string
+  isAlias?: boolean
+  aliasTargetId?: string
+}
+
+export type TabProps = {
+  color: ColorToken
+}
+
+export type DiagnosticsTabProps = {
+  overlay: string
+}
