+++ frontend/src/components/diagnostics-panel/DiagnosticsPanel.css
