+++ frontend/src/components/diagnostics-panel/index.ts
@@ -1,0 +1,14 @@
+export { default as DiagnosticsPanel } from './DiagnosticsPanel'
+export { SpacingDiagnostics } from './SpacingDiagnostics'
+export { ColorPalettePicker } from './ColorPalettePicker'
+export { OverlayPreview } from './OverlayPreview'
+export type { Props, SpacingEntry, PaletteEntry } from './types'
+export {
+  useCommonSpacings,
+  useOverlayDimensions,
+  useMatchingBoxes,
+  useAlignmentLines,
+  usePayloadInfo,
+  useScalePolygon,
+  useRenderBox,
+} from './hooks'
