+++ frontend/src/components/spacing-showcase/index.ts
@@ -1,0 +1,24 @@
+/**
+ * Central exports for spacing-showcase component
+ */
+
+export { SpacingTokenShowcase, default } from './SpacingTokenShowcase';
+export { SpacingHeader } from './SpacingHeader';
+export { StatsGrid } from './StatsGrid';
+export { ScaleVisualization } from './ScaleVisualization';
+export { FilterControls } from './FilterControls';
+export { SpacingTokenCard } from './SpacingTokenCard';
+export { TokensSection } from './TokensSection';
+
+// Hooks
+export {
+  useSpacingFiltering,
+  useClipboard,
+  useFileSelection,
+  useScaleDerivation,
+  useScaleVisualization,
+} from './hooks';
+
+// Types
+export type { SpacingToken, SpacingLibrary, SpacingTokenShowcaseProps, FilterType, SortType } from './types';
+export { styles } from './styles';
