+++ frontend/src/components/spacing-showcase/styles.ts
@@ -1,0 +1,194 @@
+/**
+ * Centralized styles for SpacingTokenShowcase component
+ */
+
+export const styles = {
+  container: {
+    fontFamily: '-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif',
+    padding: '1.5rem',
+    backgroundColor: '#f5f5f5',
+    minHeight: '100vh',
+  } as const,
+  header: {
+    backgroundColor: 'white',
+    padding: '1.5rem',
+    borderRadius: '8px',
+    marginBottom: '1.5rem',
+    boxShadow: '0 1px 3px rgba(0, 0, 0, 0.1)',
+  } as const,
+  title: {
+    fontSize: '1.75rem',
+    fontWeight: 600,
+    color: '#333',
+    marginBottom: '0.5rem',
+  } as const,
+  subtitle: {
+    color: '#666',
+    fontSize: '0.9rem',
+  } as const,
+  statsGrid: {
+    display: 'grid',
+    gridTemplateColumns: 'repeat(auto-fit, minmax(120px, 1fr))',
+    gap: '1rem',
+    backgroundColor: 'white',
+    padding: '1.5rem',
+    borderRadius: '8px',
+    marginBottom: '1.5rem',
+    boxShadow: '0 1px 3px rgba(0, 0, 0, 0.1)',
+  } as const,
+  stat: {
+    textAlign: 'center' as const,
+  } as const,
+  statValue: {
+    fontSize: '1.5rem',
+    fontWeight: 700,
+    color: '#333',
+  } as const,
+  statLabel: {
+    color: '#666',
+    fontSize: '0.75rem',
+    marginTop: '0.25rem',
+  } as const,
+  scaleSection: {
+    backgroundColor: 'white',
+    padding: '1.5rem',
+    borderRadius: '8px',
+    marginBottom: '1.5rem',
+    boxShadow: '0 1px 3px rgba(0, 0, 0, 0.1)',
+  } as const,
+  sectionTitle: {
+    fontSize: '1.25rem',
+    fontWeight: 600,
+    color: '#333',
+    marginBottom: '1rem',
+  } as const,
+  scaleVisual: {
+    display: 'flex',
+    alignItems: 'flex-end',
+    gap: '0.75rem',
+    padding: '1rem',
+    backgroundColor: '#fafafa',
+    borderRadius: '4px',
+    overflowX: 'auto' as const,
+  } as const,
+  scaleBar: {
+    display: 'flex',
+    flexDirection: 'column' as const,
+    alignItems: 'center',
+    gap: '0.5rem',
+    cursor: 'pointer',
+    transition: 'transform 0.2s',
+  } as const,
+  scaleBarFill: {
+    background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)',
+    borderRadius: '4px',
+    minWidth: '40px',
+    transition: 'transform 0.2s',
+  } as const,
+  tokensGrid: {
+    display: 'grid',
+    gridTemplateColumns: 'repeat(auto-fill, minmax(280px, 1fr))',
+    gap: '1rem',
+  } as const,
+  tokenCard: {
+    backgroundColor: 'white',
+    borderRadius: '8px',
+    overflow: 'hidden',
+    boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)',
+    border: '1px solid #e0e0e0',
+    transition: 'transform 0.2s, box-shadow 0.2s',
+    cursor: 'pointer',
+  } as const,
+  tokenVisual: {
+    padding: '1.25rem',
+    backgroundColor: '#fafafa',
+    display: 'flex',
+    flexDirection: 'column' as const,
+    alignItems: 'center',
+    gap: '0.75rem',
+  } as const,
+  spacingBox: {
+    background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)',
+    borderRadius: '4px',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    color: 'white',
+    fontWeight: 600,
+    fontSize: '0.8rem',
+  } as const,
+  tokenInfo: {
+    padding: '1rem',
+  } as const,
+  tokenName: {
+    fontWeight: 600,
+    color: '#333',
+    marginBottom: '0.25rem',
+  } as const,
+  tokenValues: {
+    fontFamily: 'monospace',
+    fontSize: '0.8rem',
+    color: '#666',
+    marginBottom: '0.5rem',
+  } as const,
+  badge: {
+    display: 'inline-block',
+    padding: '0.2rem 0.5rem',
+    borderRadius: '10px',
+    fontSize: '0.7rem',
+    marginRight: '0.5rem',
+    marginBottom: '0.5rem',
+  } as const,
+  roleBadge: {
+    backgroundColor: '#f0f0f0',
+    color: '#666',
+  } as const,
+  gridAlignedBadge: {
+    backgroundColor: '#d4edda',
+    color: '#155724',
+  } as const,
+  gridMisalignedBadge: {
+    backgroundColor: '#fff3cd',
+    color: '#856404',
+  } as const,
+  metadata: {
+    fontSize: '0.75rem',
+    color: '#999',
+    marginTop: '0.5rem',
+  } as const,
+  copyButtons: {
+    display: 'flex',
+    gap: '0.5rem',
+    marginTop: '0.75rem',
+  } as const,
+  copyButton: {
+    flex: 1,
+    padding: '0.4rem',
+    backgroundColor: 'transparent',
+    border: '1px solid #ddd',
+    borderRadius: '4px',
+    fontSize: '0.7rem',
+    cursor: 'pointer',
+    transition: 'all 0.2s',
+  } as const,
+  filterBar: {
+    display: 'flex',
+    gap: '0.5rem',
+    marginBottom: '1rem',
+    flexWrap: 'wrap' as const,
+  } as const,
+  filterButton: {
+    padding: '0.5rem 1rem',
+    border: '1px solid #ddd',
+    borderRadius: '4px',
+    backgroundColor: 'white',
+    cursor: 'pointer',
+    fontSize: '0.8rem',
+    transition: 'all 0.2s',
+  } as const,
+  filterButtonActive: {
+    backgroundColor: '#667eea',
+    color: 'white',
+    borderColor: '#667eea',
+  } as const,
+};
