+++ frontend/src/components/token-inspector/index.ts
@@ -1,0 +1,12 @@
+export { default as TokenInspector } from './TokenInspector'
+export { FilterBar } from './FilterBar'
+export { TokenList } from './TokenList'
+export { CanvasVisualization } from './CanvasVisualization'
+export type { Props, TokenRow, ColorMap } from './types'
+export {
+  useTokens,
+  useMetricById,
+  useFallbackColors,
+  useCanvasColorSampling,
+  useFilteredTokens,
+} from './hooks'
